+++ LaClock/UI/src/mods/la-clock.tsx
@@ -1,4 +1,4 @@
-import { ModuleRegistryExtend } from "cs2/modding";
+import { ModuleRegistryAppend, ModuleRegistryExtend } from "cs2/modding";
 import { bindValue, useValue } from "cs2/api";
 import { FormattedText, MarkdownRenderer } from "cs2/ui";
 import laClockStyles from "./la_clock.module.scss"
@@ -6,14 +6,15 @@ import mod from "../../mod.json"
 
 
 
+let textRenderer = new MarkdownRenderer();
+const currentSystemTimeBinding = bindValue<string>(mod.id, "CurrentSystemTime", "Error")
+const doBlinkBinding = bindValue<boolean>(mod.id, "DoBlink", true)
+
+
+
 export const LaClockComponent: ModuleRegistryExtend = (Component) => {
 
     //console.log(`Hello UI from ${mod.id}!`);
-
-    let textRenderer = new MarkdownRenderer();
-
-    const currentSystemTimeBinding = bindValue<string>(mod.id, "CurrentSystemTime", "Error")
-    const doBlinkBinding = bindValue<boolean>(mod.id, "DoBlink", true)
 
     return (props) => {
         const { children, ...otherProps } = props || {};
@@ -30,6 +31,30 @@ export const LaClockComponent: ModuleRegistryExtend = (Component) => {
                     </div>
                 </div>
                 <Component {...otherProps}>{children}</Component>
+            </>
+        );
+    };
+}
+
+
+
+export const LaClockEditorComponent: ModuleRegistryExtend = (Component) => {
+
+    return (props) => {
+        const { children, ...otherProps } = props || {};
+        const currentSystemTimeValue = useValue(currentSystemTimeBinding);
+        const doBlink = useValue(doBlinkBinding);
+
+        return (
+            <>
+                <Component {...otherProps}> {children} </Component>
+                <div className={"field_eKJ" + (doBlink ? ` ${laClockStyles.blink}` : "")} style={{ width: "auto" }}>
+                    <div className={"container_kOI container_MC3"}>
+                        <div className={"label_qsp label_mWz content_syM"}>
+                            {FormattedText({ text: currentSystemTimeValue, renderer: textRenderer })}
+                        </div>
+                    </div>
+                </div>
             </>
         );
     };